#9901: create test_utils.py  helper functions
